CR-2009-03-25-0: Bugfix:      o qconf -m* reports misleading error if tmp dir                                 is not writeable for the user and need 100% cpu                  Merged from: maintrunk (CR-2009-03-25-0)                  Bugster:     6819649                  Issue:       2960                  Review:      RD

+++ source/clients/qconf/parse_qconf.c
@@ -31,6 +31,7 @@
 /*___INFO__MARK_END__*/
 #include <string.h>
 #include <stdlib.h>
+#include <errno.h>
 #include <sys/types.h>
 #include <fcntl.h>
 #include <sys/time.h>
@@ -140,7 +141,7 @@ static int qconf_is_adminhost(sge_gdi_ctx_class_t *ctx, const char *host);
 /************************************************************************/
 
 static const char *write_attr_tmp_file(const char *name, const char *value, 
-                                       const char *delimiter);
+                                       const char *delimiter, dstring *error_message);
 
 /***************************************************************************/
 static char **sge_parser_get_next(char **arg) 
@@ -491,7 +492,6 @@ char *argv[]
 
          FREE(host);
          lFreeList(&alp);
-
          
          /* edit the template */
          argep = lFirst(arglp);
@@ -1876,7 +1876,9 @@ char *argv[]
          qconf_is_adminhost(ctx, qualified_hostname);
          qconf_is_manager(ctx, username);
 
-         centry_list_modify(ctx, &answer_list);
+         if (centry_list_modify(ctx, &answer_list) == false) {
+            sge_parse_return = 1;
+         }
          sge_parse_return |= show_answer_list(answer_list);
          lFreeList(&answer_list);
          spp++;
@@ -6890,13 +6892,20 @@ static int qconf_modify_attribute(sge_gdi_ctx_class_t *ctx,
       } else {
          sge_dstring_append_char(&delim, info_entry->instr->name_value_delimiter);
       }
-      filename = write_attr_tmp_file(name, value, sge_dstring_get_string(&delim));
-
-      *epp = spool_flatfile_read_object(alpp, info_entry->cull_descriptor, NULL,
-                                info_entry->fields, fields, true, info_entry->instr,
-                                SP_FORM_ASCII, NULL, filename);
-      unlink(filename);
-      FREE(filename);
+      {
+         dstring write_attr_tmp_file_error = DSTRING_INIT;
+         filename = write_attr_tmp_file(name, value, sge_dstring_get_string(&delim), &write_attr_tmp_file_error);
+         if (filename == NULL && sge_dstring_get_string(&write_attr_tmp_file_error) != NULL) {
+            answer_list_add_sprintf(alpp, STATUS_EDISK, ANSWER_QUALITY_ERROR, sge_dstring_get_string(&write_attr_tmp_file_error));
+         } else {
+            *epp = spool_flatfile_read_object(alpp, info_entry->cull_descriptor, NULL,
+                                      info_entry->fields, fields, true, info_entry->instr,
+                                      SP_FORM_ASCII, NULL, filename);
+            unlink(filename);
+            FREE(filename);
+         }
+         sge_dstring_free(&write_attr_tmp_file_error);
+      }
 
       /* Bugfix: Issuezilla #1005
        * Since we're writing the information from the command line to a file so
@@ -7043,20 +7052,23 @@ static int qconf_modify_attribute(sge_gdi_ctx_class_t *ctx,
 }
 
 static const char *write_attr_tmp_file(const char *name, const char *value, 
-                                       const char *delimiter)
+                                       const char *delimiter, dstring *error_message)
 {
    char *filename = (char *)malloc(sizeof(char) * SGE_PATH_MAX);
    FILE *fp = NULL;
-
+   int my_errno;
    DENTER(TOP_LAYER, "write_attr_tmp_file");
 
-   if (sge_tmpnam(filename) == NULL) {
+   if (sge_tmpnam(filename, error_message) == NULL) {
       DRETURN(NULL);
    }
 
+   errno = 0;
    fp = fopen(filename, "w");
-   
+   my_errno = errno;
+
    if (fp == NULL) {
+      sge_dstring_sprintf(error_message, MSG_ERROROPENINGFILEFORWRITING_SS, filename, strerror(my_errno));
       DRETURN(NULL);
    }
